chore: 디렉터리명 업데이트
