+++ src/constants/schedulers/major-style-notice-scheduler-constants.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the root directory or at
  * https://opensource.org/license/mit
  * Author: junho Kim
- * Latest Updated Date: 2025-03-05
+ * Latest Updated Date: 2025-03-08
  */
 
 export abstract class MajorStyleNoticeSchedulerConstants {
